pyproject.toml
[project]
name = "auth-service"
version = "0.1.0"
description = ""
authors = [
    {name = "Samandar0705", email = "samandarupwork0705@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.12,<4.0"
dependencies = ["fastapi (>=0.115.12,<0.116.0)", "uvicorn (>=0.34.2,<0.35.0)", "sqlalchemy (>=2.0.40,<3.0.0)", "pydantic[email] (>=2.11.3,<3.0.0)", "python-jose (>=3.4.0,<4.0.0)", "python-dotenv (>=1.1.0,<2.0.0)", "bcrypt (>=4.3.0,<5.0.0)", "asyncpg (>=0.30.0,<0.31.0)", "alembic (>=1.15.2,<2.0.0)", "psycopg2-binary (>=2.9.10,<3.0.0)", "redis (>=5.2.1,<6.0.0)", "pydantic-settings (>=2.9.1,<3.0.0)", "passlib (>=1.7.4,<2.0.0)", "python-multipart (>=0.0.20,<0.0.21)"]

[tool.poetry]
packages = [{include = "auth_service", from = "src"}]

[tool.poetry.dependencies]
python = ">=3.12,<4.0"
slowapi = { version = "^0.1.9", python = ">=3.7,<4.0" }

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
